rename api::parser to api::decoder
